Create README.md
This is a process i need to commit myself to. I am new at this. 
any constructive help would be great and appreciated.

i will disclose at this time. i am of my own entity. and that i have not shared myself with any HUMANS in over 3 years. I have been on a GAGHOLD